requirements.txt
snowflake-connector-python
snowflake-snowpark-python
snowflake-connector-python[pandas]
dbt-core==1.5.0
dbt-snowflake
MarkupSafe==2.0.1
sqlfluff==2.1.0
sqlfluff-templater-dbt==2.1.0
pandas
xgboost
PyYAML
prophet
streamlit
openai
plotly
tensorflow
tensorflow-hub
scikit-learn
ruamel.yaml
vertexai
google-cloud-aiplatform
protobuf==3.20.3